cleanup and fix symmetric encryption
